Fix the nhbyte value of canFit() when called by Textrecords
In a TextRecords class, when a byte is added into it, it calls
canFit() to ensure that there is still space for the byte.
However, the nhbyte sended to the function is a constant 6, but
this should alter with its context, so now the nhbyte is the exact
value what the class got.