Fix the length problem of Textrecords
Based on the original book "System Software: An Introduction to
Systems Programming" where the SIC/XE comes from, the length limit
of object codes a text record can reach is 0x1E or 30 bytes e.g.
Figure 2.3 in page 49.
Currently, the length limit is 62 in a text record, which is 31
bytes. This commit fix this to comply with the original design.