Fix split commands and formats in object and list files
This commit aims to fix commands that are split in object files.
Before that, the class TextRecords will read in the object codes
byte by byte, but this causes some commands that can't be fit in
the current record split 2 parts across different lines and recor-
ds.
Based on the original book "System Software: An Introduction to
Systems Programming" where the SIC/XE comes from, when a insturct-
ion addition to the current record will cause the record exceeds
its length limit, the assembler should take a newline instead of
split the command apart e.g. Figure 2.3 in page 49. Thus, the
addition of texts in Textrecords is currently fixed preliminarily.
Moreover, trailing space and newline in object or list files are
removed for better clarify.
Finally, the starting address following the START directive should
be seen as a hex value instead of a decimal e.g. Figure 2.2 in pa-
ge 47. This has also been fixed with all letters in uppercase.